requirements.txt
Flask>=3.0,<4.0
flask-cors>=4.0,<5.0
Pillow>=10.0,<12.0
pandas>=2.0,<3.0
scikit-learn>=1.4,<2.0
pyserial>=3.5,<4.0